moves ee test to ee folder
