+++ ee/build.gradle
@@ -25,8 +25,20 @@ dependencies {
 
     compileOnly project(":supertokens-core")
     testImplementation project(":supertokens-core")
+    testImplementation project(":supertokens-core").sourceSets.test.output
 
     compileOnly 'com.auth0:java-jwt:4.0.0'
+
+    // this is so that we can find plugin-interface jar while testing
+    testImplementation 'junit:junit:4.12'
+
+    // https://mvnrepository.com/artifact/org.mockito/mockito-core
+    testImplementation group: 'org.mockito', name: 'mockito-core', version: '3.1.0'
+
+    // https://mvnrepository.com/artifact/org.reflections/reflections
+    testImplementation group: 'org.reflections', name: 'reflections', version: '0.9.10'
+
+    testImplementation 'com.tngtech.archunit:archunit-junit4:0.22.0'
 }
 
 task copyJars(type: Copy) {
